backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-docx==1.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1
reportlab==4.1.0
pydantic==2.6.0
